Throw exception when no actions found for account
Currently an empty list is returned when an action lookup is done
on an account with no actions.  Raise a new action validation exception
when this happens instead of returning the empty list.  This provides
an indication to the client that there was something wrong.

Please see issue: https://github.com/EOSIO/eos/issues/5153